Set price as float

+++ application/core/excelservice.py
@@ -50,8 +50,8 @@ def _get_or_create_category(category_name, parent_category):
 def _create_product(product_name, parent_category,
                     category1, category2, category3, product_description, product_price, image):
     if product_price:
-        product_price = int(product_price)
+        product_price = float(product_price)
     else:
-        product_price = 0
+        product_price = 0.0
     category = _create_category(category3, category2, category1, parent_category)
     dishservice.create_dish(product_name, product_name, product_description, product_description, str(image), product_price, category.id)
